7_frontend_dashboard/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
plotly>=5.14.0
